requirements.txt
django
djangorestframework
drf-yasg
djangorestframework-simplejwt
pytelegrambotapi